login can not work

+++ routes/web.php
@@ -2,6 +2,7 @@
 
 use App\Http\Controllers\CategoriController;
 use App\Http\Controllers\CategoryController;
+use App\Http\Controllers\LoginController;
 use App\Http\Controllers\ProductController;
 use App\Http\Controllers\UserController;
 use App\Http\Controllers\UserGroupController;
@@ -18,39 +19,37 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::get('/', function () {
-    return view('layout.main');
-});
+Route::get('/login', [LoginController::class, 'login'])->name('login')->middleware('guest');
+Route::post('/authenticate', [LoginController::class, 'authenticate'])->name('authenticate')->middleware('guest');
+Route::get('/', [LoginController::class, 'home'])->name('dashboard')->middleware('auth');
+Route::get('/logout', [LoginController::class, 'logout'])->name('logout')->middleware('auth');
 
-Route::get('/groups', [UserGroupController::class, 'index'])->name('groups.index');
-Route::get('/groups/create', [UserGroupController::class, 'create'])->name('groups.create');
-Route::post('/groups/store', [UserGroupController::class, 'store'])->name('groups.store');
-Route::get('/gropus/{id}/edit', [UserGroupController::class, 'edit'])->name('groups.edit');
-Route::put('/groups/{id}', [UserGroupController::class, 'update'])->name('groups.update');
-Route::delete('/groups/{id}', [UserGroupController::class, 'destroy'])->name('groups.destroy');
+Route::get('/groups', [UserGroupController::class, 'index'])->name('groups.index')->middleware('auth');
+Route::get('/groups/create', [UserGroupController::class, 'create'])->name('groups.create')->middleware('auth');
+Route::post('/groups/store', [UserGroupController::class, 'store'])->name('groups.store')->middleware('auth');
+Route::get('/gropus/{id}/edit', [UserGroupController::class, 'edit'])->name('groups.edit')->middleware('auth');
+Route::put('/groups/{id}', [UserGroupController::class, 'update'])->name('groups.update')->middleware('auth');
+Route::delete('/groups/{id}', [UserGroupController::class, 'destroy'])->name('groups.destroy')->middleware('auth');
 
-Route::get('/users', [UserController::class, 'index'])->name('users.index');
-Route::get('/users/create', [UserController::class, 'create'])->name('users.create');
-Route::post('/users/store', [UserController::class, 'store'])->name('users.store');
-Route::get('/users/{id}/edit', [UserController::class, 'edit'])->name('users.edit');
-Route::put('/users/{id}', [UserController::class, 'update'])->name('users.update');
-Route::delete('/users/{id}', [UserController::class, 'destroy'])->name('users.destroy');
-Route::get('/users/{id}', [UserController::class, 'show'])->name('users.show');
+Route::get('/users', [UserController::class, 'index'])->name('users.index')->middleware('auth');
+Route::get('/users/create', [UserController::class, 'create'])->name('users.create')->middleware('auth');
+Route::post('/users/store', [UserController::class, 'store'])->name('users.store')->middleware('auth');
+Route::get('/users/{id}/edit', [UserController::class, 'edit'])->name('users.edit')->middleware('auth');
+Route::put('/users/{id}', [UserController::class, 'update'])->name('users.update')->middleware('auth');
+Route::delete('/users/{id}', [UserController::class, 'destroy'])->name('users.destroy')->middleware('auth');
+Route::get('/users/{id}', [UserController::class, 'show'])->name('users.show')->middleware('auth');
 
-Route::get('/category', [CategoryController::class, 'index'])->name('category.index');
-Route::get('/category/create', [CategoryController::class, 'create'])->name('category.create');
-Route::post('/category/store', [CategoryController::class, 'store'])->name('category.store');
-Route::get('/category/{id}/edit', [CategoryController::class, 'edit'])->name('category.edit');
-Route::put('/category/{id}', [CategoryController::class, 'update'])->name('category.update');
-Route::delete('/category/{id}', [CategoryController::class, 'destroy'])->name('category.destroy');
+Route::get('/category', [CategoryController::class, 'index'])->name('category.index')->middleware('auth');
+Route::get('/category/create', [CategoryController::class, 'create'])->name('category.create')->middleware('auth');
+Route::post('/category/store', [CategoryController::class, 'store'])->name('category.store')->middleware('auth');
+Route::get('/category/{id}/edit', [CategoryController::class, 'edit'])->name('category.edit')->middleware('auth');
+Route::put('/category/{id}', [CategoryController::class, 'update'])->name('category.update')->middleware('auth');
+Route::delete('/category/{id}', [CategoryController::class, 'destroy'])->name('category.destroy')->middleware('auth');
 
-Route::get('/products', [ProductController::class, 'index'])->name('products.index');
-Route::get('/products/create', [ProductController::class, 'create'])->name('products.create');
-Route::post('/products/store', [ProductController::class, 'store'])->name('products.store');
-Route::get('/products/{id}/edit', [ProductController::class, 'edit'])->name('products.edit');
-Route::put('/products/{id}', [ProductController::class, 'update'])->name('products.update');
-Route::delete('/products/{id}', [ProductController::class, 'destroy'])->name('products.destroy');
-Route::get('/products/{id}', [ProductController::class, 'show'])->name('products.show');
-
-
-
+Route::get('/products', [ProductController::class, 'index'])->name('products.index')->middleware('auth');
+Route::get('/products/create', [ProductController::class, 'create'])->name('products.create')->middleware('auth');
+Route::post('/products/store', [ProductController::class, 'store'])->name('products.store')->middleware('auth');
+Route::get('/products/{id}/edit', [ProductController::class, 'edit'])->name('products.edit')->middleware('auth');
+Route::put('/products/{id}', [ProductController::class, 'update'])->name('products.update')->middleware('auth');
+Route::delete('/products/{id}', [ProductController::class, 'destroy'])->name('products.destroy')->middleware('auth');
+Route::get('/products/{id}', [ProductController::class, 'show'])->name('products.show')->middleware('auth');
